+++ run.py
@@ -280,6 +280,26 @@ PATCHES = {
         "remove_crel.patch",
         "revert_siso.patch",
     ],
+    # android target にパッチが追加されるごとに support/android-x86_64 ブランチ用の
+    # target でもパッチを追加してビルドが通るようにする必要がある。
+    "android_x86_64": [
+        "add_deps.patch",
+        "4k.patch",
+        "revive_proxy.patch",
+        "add_license_dav1d.patch",
+        "ssl_verify_callback_with_native_handle.patch",
+        "android_webrtc_version.patch",
+        "android_fixsegv.patch",
+        "android_simulcast.patch",
+        "android_hardware_video_encoder.patch",
+        "android_proxy.patch",
+        "h265.patch",
+        "h265_android.patch",
+        "fix_perfetto.patch",
+        "fix_moved_function_call.patch",
+        "remove_crel.patch",
+        "revert_siso.patch",
+    ],
     "android_sdk": [
         "add_deps.patch",
         "4k.patch",
@@ -465,7 +485,7 @@ def get_webrtc(source_dir, patch_dir, version, target, webrtc_source_dir, no_his
         with cd(webrtc_source_dir):
             cmd(["gclient"])
             cmd(["fetch", *no_history_flag, "webrtc"])
-            if target in ("android", "android_sdk"):
+            if target in ("android", "android_sdk", "android_x86_64"):
                 with open(".gclient", "a") as f:
                     f.write("target_os = [ 'android' ]\n")
             if target in ("ios", "ios_sdk"):
@@ -702,6 +722,11 @@ WEBRTC_BUILD_TARGETS = {
         "sdk/android:libjingle_peerconnection_so",
         "sdk/android:native_api",
     ],
+    "android_x86_64": [
+        "sdk/android:libwebrtc",
+        "sdk/android:libjingle_peerconnection_so",
+        "sdk/android:native_api",
+    ],
     "android_sdk": [
         "sdk/android:libwebrtc",
         "sdk/android:libjingle_peerconnection_so",
@@ -866,8 +891,10 @@ def build_webrtc_ios_sdk(
 
 
 ANDROID_ARCHS = ["arm64-v8a"]
+ANDROID_X86_64_ARCHS = ["x86_64"]
 ANDROID_SDK_ARCHS = ["arm64-v8a"]
 ANDROID_TARGET_CPU = {
+    "x86_64": "x64",
     "armeabi-v7a": "arm",
     "arm64-v8a": "arm64",
 }
@@ -879,6 +906,7 @@ def build_webrtc_android(
     version_info: VersionInfo,
     deps_info: DepsInfo,
     extra_gn_args,
+    archs: list[str],
     webrtc_source_dir=None,
     webrtc_build_dir=None,
     debug=False,
@@ -911,7 +939,7 @@ def build_webrtc_android(
     ) as f:
         f.writelines(map(lambda x: (x + "\n").encode("utf-8"), lines))
 
-    for arch in ANDROID_ARCHS:
+    for arch in archs:
         work_dir = os.path.join(webrtc_build_dir, arch)
         if gen_force:
             rm_rf(work_dir)
@@ -1254,6 +1282,12 @@ def package_webrtc(
             dirs += [
                 os.path.join(webrtc_build_dir, arch),
             ]
+    elif target == "android_x86_64":
+        dirs = []
+        for arch in ANDROID_X86_64_ARCHS:
+            dirs += [
+                os.path.join(webrtc_build_dir, arch),
+            ]
     elif target == "android_sdk":
         dirs = []
         for arch in ANDROID_SDK_ARCHS:
@@ -1330,6 +1364,16 @@ def package_webrtc(
         ]
         for arch in ANDROID_ARCHS:
             files.append(([arch, "libwebrtc.a"], ["lib", arch, "libwebrtc.a"]))
+    elif target == "android_x86_64":
+        jar_arch = ANDROID_X86_64_ARCHS[0]
+        files = [
+            (
+                [jar_arch, "lib.java", "sdk", "android", "libwebrtc.jar"],
+                ["jar", "webrtc.jar"],
+            ),
+        ]
+        for arch in ANDROID_X86_64_ARCHS:
+            files.append(([arch, "libwebrtc.a"], ["lib", arch, "libwebrtc.a"]))
     elif target == "android_sdk":
         # M140 あたりからソースディレクトリ以下でないとエラーになるようになっているので
         # webrtc_src_dir にビルド済みバイナリが配置されている
@@ -1387,6 +1431,7 @@ TARGETS = [
     "ubuntu-24.04_armv8",
     "raspberry-pi-os_armv8",
     "android",
+    "android_x86_64",
     "android_sdk",
     "ios",
     "ios_sdk",
@@ -1422,6 +1467,7 @@ def check_target(target):
             "ubuntu-24.04_armv8",
             "raspberry-pi-os_armv8",
             "android",
+            "android_x86_64",
             "android_sdk",
         ):
             return True
@@ -1743,6 +1789,15 @@ def main():
             elif args.target == "android":
                 build_webrtc_android(
                     **build_webrtc_args,
+                    archs=ANDROID_ARCHS,
+                    gen=args.webrtc_gen,
+                    gen_force=args.webrtc_gen_force,
+                    nobuild=args.webrtc_nobuild,
+                )
+            elif args.target == "android_x86_64":
+                build_webrtc_android(
+                    **build_webrtc_args,
+                    archs=ANDROID_X86_64_ARCHS,
                     gen=args.webrtc_gen,
                     gen_force=args.webrtc_gen_force,
                     nobuild=args.webrtc_nobuild,